BigDecimal change , backgroundschedular, transactionlog class

+++ online-banking-system-backend/src/main/java/com/devteam/online_banking_system_backend/persistence/dtos/checkAccountDtos/CheckTransactionDto.java
@@ -1,12 +1,14 @@
 package com.devteam.online_banking_system_backend.persistence.dtos.checkAccountDtos;
 
+import java.math.BigDecimal;
+
 public class CheckTransactionDto
 {
     private Long checkAccountId;
-    private Double amount;
+    private BigDecimal amount;
 
     public CheckTransactionDto(){}
-    public CheckTransactionDto(Long checkAccountId, Double amount)
+    public CheckTransactionDto(Long checkAccountId, BigDecimal amount)
     {
         this.setCheckAccountId(checkAccountId);
         this.setAmount(amount);
@@ -15,6 +17,6 @@ public class CheckTransactionDto
     public Long getCheckAccountId(){return this.checkAccountId;}
     public void setCheckAccountId(Long checkAccountId){this.checkAccountId = checkAccountId;}
 
-    public Double getAmount(){return this.amount; }
-    public void setAmount(Double amount){ this.amount = amount;}
+    public BigDecimal getAmount(){return this.amount; }
+    public void setAmount(BigDecimal amount){ this.amount = amount;}
 }
