+++ online-banking-system-backend/src/main/java/com/devteam/online_banking_system_backend/persistence/dtos/savingsAccountDtos/SavingsTransactionDto.java
@@ -1,12 +1,14 @@
 package com.devteam.online_banking_system_backend.persistence.dtos.savingsAccountDtos;
 
+import java.math.BigDecimal;
+
 public class SavingsTransactionDto
 {
     private Long savingsAccountId;
-    private Double amount;
+    private BigDecimal amount;
 
     public SavingsTransactionDto(){}
-    public SavingsTransactionDto(Long savingsAccountId, Double amount)
+    public SavingsTransactionDto(Long savingsAccountId, BigDecimal amount)
     {
         this.setSavingsAccountId(savingsAccountId);
         this.setAmount(amount);
@@ -15,7 +17,7 @@ public class SavingsTransactionDto
     public Long getSavingsAccountId(){return this.savingsAccountId;}
     public void setSavingsAccountId(Long savingsAccountId){this.savingsAccountId = savingsAccountId;}
 
-    public Double getAmount(){return this.amount;}
-    public void setAmount(Double amount) {this.amount = amount;}
+    public BigDecimal getAmount(){return this.amount;}
+    public void setAmount(BigDecimal amount) {this.amount = amount;}
 
 }
